inventory_service/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy==1.4.*
pydantic
python-jose[cryptography]
python-dotenv
requests
